Remove Google Analytics
It's easier if we don't have to explain
what's being logged, why it's being logged
and why users (who might be pasting sensitive
requests) should be okay with it.

To be clear, it was being used to collect
statistics about which languages were
being converted to and if the conversions
were succeeding. The actual pasted curl
commands (or their conversion) were not logged.